+++ trivia/trivia/GameResults.cpp
@@ -4,9 +4,9 @@ GameResults::GameResults(std::string username, unsigned int correctAnswerCount, 
 {
 }
 
-int GameResults::score() const
+float GameResults::score() const
 {
-	return this->_correctAnswerCount / this->_wrongAnswerCount / (this->_averageAnswerTime + HIGHSCORE_ANSWER_TIME_OFFSET);
+	return (float)this->_correctAnswerCount / this->_wrongAnswerCount / (this->_averageAnswerTime + HIGHSCORE_ANSWER_TIME_OFFSET);
 }
 
 bool GameResults::operator<(const GameResults& other) const
